move space membership allowed auth policy extension to community auth service
